Add logic for querying by date
We can now get pdbs given a date range. This is needed for the pipeline since we
will be rerunning the nr set many times. Each time will be for structures before
a date. This will help me run these structures easily and concisely. Also this
comes with tests about things so I know that the structures are gotten are
actually correct.