Add date options and organize options
We want to be able to run the pipeline using structures with a given date range.
We use the recently added options in the rna helper to do so. This allows us to
more easily rerun the pipeline on past inputs. I also have organized the many
options into more logical groups. In addition I changed the priority of some
options. This way all takes priority over all other options.