Prepare for generating ids with alt ids
Some unit ids use alt ids. In order to get correct quality information for these
we need to use the alt id when we generate ids. I haven't fully tested this yet
but we do have some code for it. In addition, I've found a pdb that this can be
tested on so once we bootstrap it we can complete these tests.