Optimize query
The query worked fine for small inputs. However, with large inputs it is too
slow. This modifies it be be faster by making fewer joins, correspondence_pdbs
had several joins in it, and using indexed fields. It may still have problems
with a maximum number of arguments for when the number of pdbs gets too large.
This can be dealt with later I think though.