Improve generation of pairs to test
This improves the generation of pairs. This way we do not create all possible
pairs and then filter them. This can produce a very large number of pairs to
compare. Instead we generate only the things that have alignments. This cuts
down drastically the number of pairs. We could reduce it further by the number
of good alignments, but that isn't needed yet.

I've also improved the testing of this grouper by adding more basic tests. I
will work on adding more tests that cover more of the complex behavior.