Ensure all pairs are possible
I don' think this should happen but we should filter all pairs such that they
have ids which do actually exist in the mapping. This worries me a bit, as I
feel it shouldn't be needed, but it is a good idea. I may come back later to try
to figure out why this happens.