Try merging ife stages
The ife stages are a bit special they end up recomputing the same data twice,
which is slow. It would be better to do it once and then not have to do this
again. We can do this by either caching or storing everything at once. If we
used caching it would create thousands of files that aren't really needed. It is
better to just store everything at once and be done with it here. There is
little extra to compute, unlike the nr stages.