Move all setup logic to it's own module
Frankly it was getting long and hard to follow for me. This way it is all in one
place. Ideally, I could also test it. But at least like this it is a bit
simpler to track everything that needs to happen during setup. I would also like
the matlab error logging to be handled as well. That will need to be fixed in
later versions.