Improve species detection
We improve the species detection to handle the common error case of a chain be
named as from ecoli but labeled as species 512, instead of 562. There are likely
other error cases but we only handle this one for now. We could also check if
the name differs from the assigned species name. That poses it's own problems
when the name is an organism, but may be a good extension.

We also have to remove one test for having more than one species and it
assigning synthetic. The data no longer has more than one species so the test
has to go.

I have also changed the habit of marking everything synthetic when we have no
species. I don't like that behavior as it confuses what the correct species is.