Correct logic for opening files to save to
We would open files in the wrong mode sometimes. Notably we would always open
for writing or always for appending. This is not what should happen. What should
happen is that if merge is not set we should open for writing on the first
chunk, otherwise we should open for appending.