Update ife generation logic
We now use some updated logic for the generation of ife's. This makes the logic
match what we believe it should do.

In short the logic is we begin by marking chains as structured or unstructured.
We always connect any unstructured chain to any other chain it interacts with.
We only consider cWW basepairs. We connect structured to structured only if it
makes enough cWW between them. When connecting a structured to an unstructured,
that is already connected to a structured we do not put the two structured in
the same group. Instead we put the unstructured ones into both groups, but do
not connect them.

Not all interesting cases have been tested yet but this is enough to work with.
It passes all cases I have tested. We will need to find more examples for more
complete testing.

I have also added a module to store constants. This will be where I keep the
constants needed for cutoffs and the like. I'd like to move all constants and
the like into there.