Use a newer setup for storing ifes
I've moved ife generation to create objects other than dictionaries. This needs
to be reflected in how we compute the data. In addition, we also move storing
the ife and chain level stuff into one step so we don't have to recompute data.
This will speed up the pipeline a fair bit I hope. Plus it is silly to recompute
stuff that doesn't need to be recomputed.