First attempt at using a custom compiler extension
We can make delete statements in mysql work by adding a compiler extension. This
changes how the sql is generated. Hopefully this will produce the correct sql
that deletes only the correct tables. As this is code may have some large
effects on the pipeline I will wait to use it until this has been properly
tested. However, I have other more pressing issues now so I will get back to
this later. We will have to test the string generation as well as testing out
the effects in the database to make sure everything works.