Yet another rewrite of this module
I have added logic to only use the normalized sequences and data. This should
make sure we never create correspondences of 0 length. That is a key part of
normalizing sequences.

I have also completely rewritten the logic again. This attempt should combine my
first two attempts. It does everything in python to make it more testable and
understandable. It shouldn't be too slow as it hits the DB as few times as
possible and it computes everything at once. This should make it the easiest to
understand and maintain of all other approaches. Hopefully at least. This isn't
yet tested, but will be later.