Initial work on importing 2d structures
We need to import secondary structure data for a nice display of 2d diagrams. We
will import them using the data provided by the gutell lab for doing this. I
have started this with a very simple and untested pair importers for the data. I
have also copied over what I think are my old parsers for this data. It all
appears reasonable but I need to test this out. These importers may also be
merged into one as they parse the same file twice. This is kinda unneeded so I
may fix that after testing things out.

This is also the first of a couple steps for adding secondary structure data. We
will also need to figure out what sequences to align the secondary structure
data to and then do the alignments.