Add first version of ss data import
We can now import a secondary structure diagram to the database. This is a
pretty simple set of changes to make it possible. Most of the work was done a
while ago in creating the parser. I have added a command to the main script to
run the importing steps. This does not yet align but that is the next step.