Allow mapping of experimental sequences to 2ds
This adds the logic required to map the positions in an experimental sequence to
a 2D. I haven't tested the ability to infer alignments since that is more
complicated and relies upon nr sets working. For that reason I have removed the
dependency of nr stuff from the ss stuff.

For now it is only possible to do this alignments by manually putting in a
structure and chain to align on the command line.