Start moving to using click
I want to use click to build up the command line interface as it is more
flexible and a bit easier to do. Plus the options seem a little more reasonable
so long as you aren't using the advanced stuff. This is a start to that work.
This should make things like documenting the commands and listing all possible
stages easier.