Reorganize matlab
The definition of matlab in the core causes a variety of annoyances. Mostly,
since it is possible that matlab is not around the import will fail. In this
case we should log things. However, the logging may not already be setup unless
we delay loading things until very late. For example of this look how we have to
load the dispatcher module within a function. If I just move Matlab out of the
core this isn't needed. Plus it isn't really a core thing anyway. Mostly it is
just a utility class, so it does make sense there.