Improve the cli interface
This improves the CLI by moving more logic to where it belongs in the cli part.
The setup module is much reduced in size, and could possibly be removed. This
interface needs some more testing before it can be finalized. In addition, I
would like to make it automatically know all valid stages to run. That would
make it easier for new people to start working with it as the valid stages would
be listed, possibly with some documentation.