Begin some work on parent counts
This starts the work on creating the loader for parent counts. This stage isn't
tested yet as I don't have several releases as part of the bootstrapping
process. I will have to test it by running it on various databases and checking
out what happens. I will add bootstrapping to the database eventually.