Add list and about commands
We can now list out nearly all known stages as well information about a specific
stage. This is based on introspecting the pipeline. It is not perfect but does a
pretty good job. Stages which are not part of pymotifs.update will not be found.

This also highlights the need for documenting all parts as the documentation is
now a part of what is listed to the user. I was planning on doing this but this
makes it much more urgent.