Add more functions to introspection
It will be helpful to move all the introspection into one module. Some logic,
like finding a loader from the name is duplicated. It exists within the
dispatcher but will be moved into this module as well. For now it is untested
and unused. I will have to test it before using it.