Fix issue with ifes not being generated correctly
Some ifes were not built correctly. Notably, for 1YZ9 we end up producing more
IFE's than we should. This is because we link 2 ifes but do not propagate the
linking to all ifes that are already part of the links. So we can join C to D
and then D to E but not note that C should be with E. Thus, we must make sure
that all chains which have been joined into a ife are marked as part of the ife.
This should fix the issue. I'm pretty sure this won't in inadvertently link a
structured to an unstructured here.