Rewrite species mapping logic
We are having troubles with the xml modules that biopython uses on rnatest.
Because of this I have rewritten this to not use biopython. It is also now
pointed at a European resource since they make no claims on how often and how
much data you can request from them. I doubt it will be an issue since we will
generally only need a few new species ids each time we get new structures. We
also now handle getting genus or above level data by inserting NULL to the
columns to indicate no known species for it.