Deal with oddly formatted taxonomy ids
Some chains do not have a single taxonomy id. For example 4FWT|A has a composite
id of 562#32630#39803. I'm not going to do anything with this other than pretend
it is synthetic or no species known. Thus we won't try to map this to a species
and instead just not add it to the list of species to process. We do however log
this so we can see it in log files. If it becomes annoying we can do something
about it.