Test and log for reflexive counts that are unequal
This seems to occur on the development server to 2MKN, however, I can not
reproduce the issue on my machine. As the difference seems to be small I will
just log it as a warning and use the higher connection. Hopefully, I'll be able
to track down the cause of this if I have more examples. Or I could recompute
all interactions on development for this structure.