Update id generation of ife's
We should only list the structured chains in an ife id, unless there are no
structured chains, in which case we list the first chain. The changes here all
relate to that.

We begin by correcting the __lt__ method so it works correctly on ife chains.
This allows us to correctly order the chains for selecting what goes in ids. We
also can get rid of the __gt__ method with this.

We also add a slot for data about bps and make __eq__ and __lt__ use the same
data.

We then move on to correcting how the ids are generated. This requires changes
to many tests to reflect the new naming scheme.