Add option to download pdb files
The option is probably not named correctly. This will download the pdb files,
but somethings like fr3d in matlab will use a cif over a pdb file. Not much to
do about this yet I think. However, it is a start and hopefully will deal with
the immediate problem of using cif files instead of pdb files in the nr set.