Changes to respect models and symmetry operators
When loading IFE's we need to respect the model and symmetry operator data.
Because we don't do this we end up with lengths and base pair counts that are
incorrect. This is now fixed. We load a symmetry operator, doesn't matter which,
and the model with the most basepairs to load the ife's and interactions.