Add logging and error catching
There was a problem where not all ife's and structures were being clustered. To
find and fix this I used the logging and error handling here. I've also reverted
the pairs logic to the older slower method. I will try the faster version and
hopefully that can work once we ensure we always do alignments before using it
again.