Updates to deal with chains that have rna
Some chains are listed as RNA but they contain no RNA bases. This can be due to
modified bases or possibly a mistake in PDB's web service. To deal with this we
need to filter out all chains that have length 0 after we load IFE's. I've added
code for this as well as some corrections to tests and such for the nr builder.