Always group ifes with themself
We should make sure that the graph of ife connections includes the self
connections. This can be a problem if an ife has no connected pairs. In this
case it will never be inserted into the graph. It will then not be part of any
produced group. This then leads to an issue where not all ife are grouped. We
can avert this by adding the self connection so even singleton groups can be
produced.