Update and test the correspondence utilities
These had gotten out of date and were untested. I've added some simple tests for
the as well as reworking the code to use sql alchemy query building instead of
just string interpolation. This required a change to the models to add primary
keys for the correspondence_units table.

I have removed the pdb_ordering method as it is not well thought out, there is
more than 1 correspondence between pdbs. It is also not used in any active code.
There is a reference to it in correspondence.interactions but that isn't used
and needs to be updated anyway.