Make bootstrap command use bootstrap config
This forces it to always use the bootstrap config. This might not be
ideal since we could want to use a different one. However, this does
prevent what I've found as the ore common error of forgetting to not use
the standard configuration. Forgetting to do so will stomp over test
data and exploratory work.