Add some utility classes for temp tables
Sometimes when querying it is useful to build a table that contains some
temporary information. This handles building, querying and removing such
tables. This issue arises because you can only send so many parameters
to a sqlachemly query. When using all PDB's we go over that number. So
instead we can build a temp table to query off of. These helpers make
that much more convenient.