Test representative finding logic
It seems we can more or less find the correct representatives. The one
caveat is that we will change very often for small chains. For example,
a small chain with 2 may be a representative, but if we get an example
with 3 it will switch. However, we won't change much for larger
structures. I think this is ok. Right now there is a test which
displays this behavior that is marked as an expected failure.