Add an sql dump for testing
Properly testing this requires a large database setup. Computing
everything takes at least 12 hours and matlab. To make it possible to
test this on travis we need to add a dump. This file can be used to
build a database quickly and then run tests as needed. If I want to run
tests on travis this is required.