Put off having a stage fail
It is more useful if a stage waits until all data has been processed to
fail. This makes it easier to process data in the face of errors in that
we may get many different parts done before giving up. For example, the
unit ids used to not have alt ids included. Now that they do there are
several stages that complain about bad fk constraints. We can find them
all at once as well as complete all the stages that do work by running
this way.

The exception to always continuing is when the remove call triggered by
a failed entry raises. In this case we have to stop as we may have a
database in some inconsistent state that a person will have to fix
manually. This forces remove to never raise an exception if it is well
behaved. I don't think all implementations do this yet. It is possible
to check if we are in an inconsistent state by adding a call to has_data
to check but this is more coding and adds another place exceptions may
arise. I may add it in future commits.

Once we have processed all stages we will fail with a message indicated
all failed inputs for a human to deal with later.