Semi working version of cc with new fr3d
The new fr3d version allows for comparing using only matrices, or will
once that code is tested and working. This sets up most of the logic
needed to work with that. It can load the required data and send it to
be computed. Right now discrepancy is not tested since fr3d doesn't work
yet. However we have covered the rest of the methods, hopefully.