Correct testing/usage of matrix_discrepancy
We can now use the improved version of discrepancy calculations
correctly. This should take much less time and be much easier to compute
for the nr sets as we do not read in a whole cif file each time, but
instead just query the database. This activates some tests that were
skipped while the matrix_discrepancy was being implemented. We also
correct the order of arguments to it so the code now works.