Update the dispatcher and toposort
I found a problem with the previous dispatcher logic where it could not
deal with skipping dependencies of a StageContainer. It also would not
always order the stages correctly. I've reworked both the toposort and
the dispatcher to deal with this.

The toposort is taken from a python package with some minor formatting
changes. It also allows access to individual levels to run. This is
useful for future work in the dispatcher.

The dispatcher has a new more testable design. The logic behind skipping
and allowing is probably more complicated than needed, but works for
now. I also allow for access to levels to run. This may be useful for
future work to parallelize parts of the pipeline, as each stage at a
given level may be run independently. It also helps me split the logic
up into some more testable units.