Correct logic for bootstrapping
This makes sure we override the default config data with what is used by
bootstrapping. We also correctly define what stages to skip.
pdbs.obsolete must be skipped since it will load obsolete pdb_ids which
are not in the pdb_info table. This causes fk issues and so must be
skipped when bootstrapping.