Let StageContainers build the stages they contain
This allows a stage container to build all stages it contains. We store the
arguments used to build the container, as well as adding a new method which
uses those arguments on each stage. It will return the built stages in a random
order, as the defined stages are often a set.