Update logging to include for rep changes
We should clearly log when the representative changes. I add a statement
that covers that. It is sort of covered with the debug level statements,
but those aren't used by default. This will be. It will also only cover
the cases we mostly care about, when rep changes due to our new rule.