Speed up finding pairs to compare
Previous logic would hit the database many, many times to load all
pairs. This is not good as it can take way too long to find what to do.
We now load up all good correspondences ahead of time and then filter
the possible pairs. This should be much faster.