Correct bug with dependencies on StageContainers
This did not work correctly as some stages, like ife.info, did not have
their dependencies correctly computed. This fixes the issue by allowing
stages to depend on stage containers. It requires the flatten utility
method because we form a potentially very nested list of the stages that
each stage depends on.