Initial pass at storing coordinates in cif format
We use the cifatom writer in fr3d to write cif formatted text to the
database. We cleanup the data a little bit by removing lines that aren't
wanted on a per residue basis. We also do not write waters as we do not
generally view them and thus reduce the space used a little bit.