Attempt to improve recalculation detection
We don't have to recalculate data if the stage did not produce data and
was run. This tries to add that data by checking if we have no data and
we are allowed to produce no data, but we have already marked the stage
as done. If that happens then we don't recalculate.