Tests and working version of interaction summary
This tests out the logic behind creating a summary of the number of
interactions for all rna bases in a structure. This is a very useful
computation for either analysis or fr3d searches. This is a first pass
at computing such data. It relies upon the database setting missing
values to 0.