Correct loop quality steps
As far as I can tell we now correctly detect loop quality. This includes the
complementary sequence logic that was not working before. It is tested on
several structures of varying sizes and produces nearly identical results as
before. The changes appear to be due to a more senestive detection of residues
with 0 occupancy.