Attempt to use python to get aligned chains
Getting aligned chains for several thousand structures is very slow (hours).
I'm reimplementing the logic in python so hopefully it can run faster. Right
now I do have to jump through several hoops to get the same output. Hopefully
this will be faster but it still needs to be tested.