Move and add tests
I moved the namer tests a bit so they are in a better place. I've also
added some test for the newly added change counts class.

This class should allow me to correctly get the parent counts. The
simple tests show that it works as expected on simple data and it will
be used in the nr and motif builders to create the required counts.