Use method when picking representative
When we pick a representative we should also consider the method used.
Right now we think that x-ray is always better than cryo-em. This is
because the cyro-em are built using x-ray models very literally. They do
not remodel extensively. This is not ideal for our motif analysis so we
always use a cyro-em over x-ray.