Try to fix ordering issue
Some loops are not named properly due to an ordering issue with old unit
ids. The problem was that when I translated old to new ids we did not
resort the ids. This means that the when we try to re-extract a loop we
end up with unit ids in a different order than in the database. This
means we give the loops a new name even though we should not. Hopefully
by sorting the unit ids used both in the loop file and the database we
should prevent this from happening.