Add a class to correct bad unit ids
When we updated the database to new style ids, some old style ids were
not translated properly. Most of the issues are due to the model number
no being the correct one. This adds some logic for deal with that. Also,
it is likely that some units will have issues with alt ids.