Add an NR report
I've had to run this several times for Dr. Zirbel and now Neocles so I
figure I might as well put the code in the pipeline for ease of use. The
idea is to have a script which can produce a simple file that can be
put in a spread sheet to examine how well the NR set building went.

I'm not sure I like how this code is organized and written right now but
it does work. Untested but the output looks reasonable.