Add another report
It will be useful to know when some species assignments are possibly
incorrect. I wrote such a report. This prompted a cleanup of the
reporting code. The species report is currently untested and has known
issues.

1. It is too sensitive to conflicting species assignments. It will
   complain even the assignment is reasonable (tRNA with > 1 species is
   ok).
2. It has no way to store and respect that some assignments are known to
   be ok.

These will be fixed later, this is just a first pass at such
functionality.