Basic correct logic working
We have a basic correction for units working. This works on sane test
data, but hasn't seen extensive testing yet. I've also done some
cleaning on the logic in the corrector. Now the mapping is used
automatically, so there is no need to use it outside the corrector. This
breaks loops.extractor and loops.positions.