Fix some issues
The chain_chain logic now needs to depend on the sequence only
groupings. This is because we will be sure to compare all chains within
an equivalence class. The logic to do so is not yet implemented, just a
few cleanups for dependencies and respecting resolution issues.

In addition, I have fixed some naming issues in constants (poor spelling
and poor choices).