Strip out old error checking code and fix comment
The comment on the possible change is now slightly better.

We also remove the old error checking code. This code will no longer
work because we aim to only load the good alignments. This means we will
not have a full matrix so the logic of checking to see if there are
pairs not in the alignment matrix will always fail. For the moment I
leave the method for this check in place. It could be useful later with
a more compact representation of the alignments.