First attempt at nr pairs report
We need to know information about all pairs of ife's in all NR groups.
This is useful information for figuring out cutoffs. To this end I've
added a report to generate this information. This is the first untested
attempt. It is basically just a port of the SQL query I've used in the
past for this.