Add report to compute chains in representatives
It is useful to know what the representative would have been for a range
of releases. This helps use evaluate the stability of various methods. I
have added a report to compute this.

In order to do this I refactored the representative finder so that it
looks up the class to use and then dispatches to that class. I have
added a module to contain various representative files. In addition,
I'ved a simple finder that only looks at bp/nt.