Improve representative selection
The representative selection should have been iterative. As in, it
should select the best bp/nt and then look for all chains that are
longer or have more bp. It should then find the chain with more bps and
nts despite a worse bp/nt with at least some percent increase. This
selection process should iterate until it has settled onto a single ife.

In addition, it also useful to compare against something a similar
method that does not require a minimum percent increase. Thus we
refactor where the percent increase parameters are given and add another
selector class. Plus we alter the names to reflect the functionalities.