Add tests specific to representatives and method
I've added some tests that are specific to the representative selection
methods. These are just modifications of the previous tests as the
representative classes are just a reorganization of the previous setup.

In addition, I've added a new method of selecting the representative.
This method is based off the percent increase method but uses the parent
the representative if it exists. This should be more stable, in theory
anyway. I haven't yet modified things to load the parent representative
though.