Correct parent assignments for nr classes
The nr classes have an incorrect parent assignment after filtering. We
determine parentage to determine names and this is based off the 'all'
resolution cutoff. While this has really nice properties of making the
handles consitent between resolutions it does cause other problems.

Notable we do not store the parentage to a group with similar
resolution. Also, when we need to pick a representative we *must* use a
group of similar resolution to ensure that the initial representative is
from a correct group. Imagine if this did not and we tried using a
representative that exists at the 'all' resolution but not at say 1.0.
This is not ideal. This should fix these issues.