First version of new quality importers
We probably need a more complex quality import because the uses for
quality data are increasing and getting more complex. To that end I've
added some new configuration as well as some more complex importers.
Notably, we will need structure level information so I'm adding an
importer for that. To make some importing a bit cleaner, as well as a
bit better about not reattempting stuff it should not, I've added a
downloader that will download the reports and store them. I may need to
add something to get chain level information as well.