Begin work on an improved nr report
We need to include things like bp/nt in the report. I've added this so
the report should be more useful now. I will also need to add other
measures of quality that will be used as part of the representative
selection. For that reason this is not complete, but it is a bit
improved and I will build off it.