Begin work on selecting representatives by quality
We have to use quality metrics to select the representative as this will
make the motifs much better. This is as per a discussion with Helen,
John and NDB/PDB. This is my start on such a method. It doesn't work yet
but this is a general outline of what I plan to do.