A pass at tools for altering dicts
I'm sure there is a much more elegant functional language way to do
this, but I don't know of something in python that does it. This is a
basic set of tools for renaming parts of a dict as well as altering the
types of values. This is a task I end up doing a lot, particually with
the quality XML reports. Hopefully this will make that logic easier to
implement.