Change error to skip
This is required because it is possible for there to be no geometric
data to be loaded due to missing nucleotides. For example, 2ZM5 chain D
and 1VY7 chain AY have identical sequences and both do have some
geometric data. However, the nucleotides that they have data for are
disjoint. That is 1VY7 only has data for the 3' end of the molecule. On
the other hand 2ZM5 chain D only has everything up to the 3'. Thus while
they align exactly and have centers/rotations it is not possible to get
aligned centers/rotations. I suspect this fairly rare, however there are
several failures in the current run of the pipeline due to this. Thus it
is ok, probably, to change the error into a skip. There may be a better
way of detecting this, but I'm not sure right now.