Some improvements to exp_seq.mapping
This alters the logic of the mapping so that the exp_seq.mapping loader
relies more on the data produced from exp_seq.chain_mapping. This is a
much more sane setup, but is very different from the previous approach.
I have also used the SimpleLoader base class to get several useful
behaviors.