Initial work to fix bug due to odd labeling
Some chains (2HOS A, 5J4D A/FB/HA/MC) should probably be labeled as
'Polyribonucleotide (RNA)' as they are RNA but are labeled as 'DNA/RNA
Hybrid'. This is likely a mistake in PDB which someone should contact
them about and fix. However, on our side we can allow these chains to be
considered RNA. This will alter what get selected for experimental
sequences and alignments. This change is a large effect so it defaults
to off, while it is tested out.