Attempt to do all stages
This may allow the pipeline to compute all data across all stages,
despite stages failing. The idea is to detect what ids caused an issue
and then simply remove those from future processing. This may very well
be a bad idea, as it could be possible to put the database in a weird
state. I'm not sure how this weird state would happen but it may be able
to. However, the data does need to be computed so this may at least do
that. If it doesn't this is the start of work to allow it.

We should probably add more logging to note that this is remove ids from
processing, but each error is logged individually at least. Having them
all in one place may be nice though.