Alter builder not re order members
The representative finders are now expected to re order members of the
eq class and put the representative first. This is a change from the
previous design but is better suited to future work. Previously we
always reordered based upon bp/nt.

We are planning on reordering and selecting members of an EC based upon
quality metrics. This change allows that to work.