Implement initial selection by quality
This is the first implementation of the logic of selecting by quality.
The ideas are Dr. Zirbel's and the basic idea is to weight by the fraction
of rsrz and backbone outliers, clashscore and resolution. We also bias
toward longer structures with more basepairs.

This is a simple implementation that may not do the 'right' thing but
does at least run. We will build off this to see what the right solution
is. For now this can be tested out on various structures to see if we
like what it does.