Attempt to fix connection leakage
When running units.distances on all structures it can use up all DB
connections. This error doesn't have a clear cause but seems to occur
when we have lots of structures already computed. I am unclear as to why
this happens but an attempt at fixing it is to only compute for new
structures of which there are generally very few.