Add stage to add motif annotations
These annotations are a mix of generated data, like the base pair
signature and manual like the 'annotation' column. This adds some logic
for importing the generated data and propagating the manual data. It
will only propagate the manual annotations if, all parents are annotated
with the same annotations. This is pretty rare and probably only happens
in cases with a single parent.