Attempt to fix issue with multiple models
Some new(?) validation reports include a model number this model number
prevents the correct mapping of unit ids to quality data. To deal with
this we attempt to fetch the model number of possible. If we can't we
assume it is 1.

This will be an issue for structures with many models as we may end up
only having quality data for model 1 which is not the model of interest.
This can probably be fixed later though.