Reorganize and implement new representative method
This moves the representative selection code around. It was getting very
large so I figured multiple files would be easier for me to work with.
At the moment this breaks the functionality of representative selection.
Basically I separated out the methods by how they work. I suspect we
will delete the methods that use structure data and keep only the
quality ones, but until that is tested I won't do so.

In addition, I've added a first pass at implementing the compscore
method for selecting the representative. This isn't yet tested. I've
also built a small base class for the quality methods so I can easily
have some common behaviors.