Improve clash data handling
It is possible, in theory, to have a structure with 0 clashes. We need
to distinguish this from a structure where not clash analysis was done.
To do this, I try looking at the global clashscore to see if it exists.
If not then the structure has never been analyzed for clashes and so the
pipeline will store NULL values for the data. If the clashscore is
present then the computed values will be stored, which may be 0.