Allow no data in quality.units
It is possible that some structures will have no data to write. For
example 1gsg does not have any data to write.

It would be better to detect this case and raise a Skip if we can be
sure that there is no data. I'm not sure how to do that.