Implement opening of compressed mmCIF files
This should allow the pipeline to read compressed mmcif files. This will
help with the issue of disk space being used up. This will read a
compressed file if it exists, and fall back to uncompressed if required.
It then fails if not.

In addition, I removed unneeded helper classes. These weren't used
anywhere other than to load mmCIF files so they are too complex.