Implement fetching structures by date range
This allows the fetching of structures using a date range. This does not
support any other options other than dates now. This is ok because that
is all that is used. We also assert that there are no obsolete ids are
fetched as well. This seems to work well now.