Create table to store clashes
Computing the compscore method is complex enough that we need a table to
store the clash information. Doing so lets me work with the desired data
easily. Doing this means create a new stage to load the data as well a
new part of the parser to build it and remove the clash data entries
from unit_quality.