Filter against units without chain_index Error observed in 1A34|1|B+1A34|1|C where the number of observed nucleotides was greater than the experimental/maximum number, producing a negative fraction_unobserved and causing a pipeline error.
Working between tables unit_info and exp_seq_info, the difference
maps to unit #11 in chain B, which is the only unit in the IFE
to not have a chain_index value.

Addition of this filter allowed a test run of nr groups report
generation to complete (pending review for accuracy).