Update rna queries to use specific units The original form of these queries uses the unit_type_id to identify RNA nucleotides.  This worked because only the canonical ACGU nucleotides were marked with unit_type_id = 'rna'.
However, for the presentation of circular secondary structure
diagrams, this practise means that any position containing a
modified nucleotide would not be shown (in fact, a silent
failure).  Not a desirable behavior.

This changes alters the query logic to request the specific units
of interest (i.e., unit IN [A, C, G, U]) instead.  (Note that the
original unit_type_id constraint is retained, as it is still
correct and might help with filtering data in certain scenarios.)

Also note that, at present, updating modified nucleotides to have
unit_type_id = 'rna' is a manual process.  A pipeline-level change
that will automate the process is under design.